+++ src/sagemaker/sklearn/processing.py
@@ -43,7 +43,6 @@ class SKLearnProcessor(FrameworkProcessor):
     def __init__(
         self,
         framework_version,  # New arg
-        s3_prefix,  # New arg
         role,
         instance_count,
         instance_type,
@@ -52,6 +51,7 @@ class SKLearnProcessor(FrameworkProcessor):
         volume_size_in_gb=30,
         volume_kms_key=None,
         output_kms_key=None,
+        code_location=None,  # New arg
         max_runtime_in_seconds=None,
         base_job_name=None,
         sagemaker_session=None,
@@ -63,7 +63,6 @@ class SKLearnProcessor(FrameworkProcessor):
         super().__init__(
             self.estimator_cls,
             framework_version,
-            s3_prefix,
             role,
             instance_count,
             instance_type,
@@ -72,6 +71,7 @@ class SKLearnProcessor(FrameworkProcessor):
             volume_size_in_gb,
             volume_kms_key,
             output_kms_key,
+            code_location,
             max_runtime_in_seconds,
             base_job_name,
             sagemaker_session,