+++ tests/unit/sagemaker/model/test_framework_model.py
@@ -56,7 +56,7 @@ class DummyFrameworkModel(FrameworkModel):
             ROLE,
             ENTRY_POINT,
             sagemaker_session=sagemaker_session,
-            **kwargs
+            **kwargs,
         )
 
     def create_predictor(self, endpoint_name):
@@ -71,7 +71,7 @@ class DummyFrameworkModelForGit(FrameworkModel):
             ROLE,
             entry_point=entry_point,
             sagemaker_session=sagemaker_session,
-            **kwargs
+            **kwargs,
         )
 
     def create_predictor(self, endpoint_name):